% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSpeciesProportions.r
\docType{data}
\name{AllSpeciesProportions}
\alias{AllSpeciesProportions}
\title{Proportion of species straddling Georges Bank EPU}
\format{
A data frame with 4 columns

\describe{
\item{AREA}{Statistical Area code}
\item{NESPP3}{species code from commerical fishing database}
\item{InOut}{Designation of insie or outside Georges Bank}
\item{MeanProp}{Mean proportion of landings within AREA for each species}

}
}
\usage{
AllSpeciesProportions
}
\description{
Georges Bank EPU is defined as an area within a subset of predefined Statistical fishing areas.
Landings from a statistical area were \href{https://noaa-edab.github.io/ms-keyrun/allocateLandingsEPU.html}{attributed to Georges Bank} by determining how variable landings were over time and if constant assign a
fixed proportion of the landings to the Georges Bank region
}
\keyword{datasets}
