% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchIndex.r
\docType{data}
\name{catchIndex}
\alias{catchIndex}
\title{The fishery catch indices for the mskeyrun project}
\format{
A data frame

\describe{
\item{NESPP3}{Northeast Species Identification code (see \code{focalSpecies})}
\item{YEAR}{year landed}
\item{variable}{commercial landings or commercial discards}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
catchIndex
}
\description{
Fisheries landings data for Georges Bank
}
\keyword{datasets}
