% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheryLenSampN.r
\docType{data}
\name{fisheryLenSampN}
\alias{fisheryLenSampN}
\title{Fishery sampled number of lengths for input into hydra model}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 455 rows and 5 columns.
}
\usage{
fisheryLenSampN
}
\description{
Number of fish measured by year, species
for mskeyrun focal species
}
\details{
#' @format A data frame

\describe{
\item{YEAR}{year diet data collected}
\item{SVSPP}{SVSPP code for species}
\item{lensampsize}{number of fish measured for length}
\item{SPECIES_ITIS}{SPECIES_ITIS code for species}
\item{LongName}{Common name for species}

}
}
\keyword{datasets}
