% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleets.r
\docType{data}
\name{fleets}
\alias{fleets}
\title{Fleet designations based on NEGEAR2 codes}
\format{
A data frame with 3 columns

\describe{
\item{fleetName}{Name of fleet}
\item{NEGEAR2}{Gear codes associated with each fleet}
\item{hydraFleets}{Fleet aggregation to satify Hydra multispecies model}

}
}
\usage{
fleets
}
\description{
NEGEAR2 codes are assigned to fleets based on which species are landed.
The methodology and the decisions made to create this data set can be found below
}
\section{Methodology}{


This data set was created based on clustering \href{https://noaa-edab.github.io/ms-keyrun/GBFleetDefinitions.html}{methods}.
}

\keyword{datasets}
