% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalSpecies.r
\docType{data}
\name{focalSpecies}
\alias{focalSpecies}
\title{The focal species in mskeyrun project}
\format{
A data frame

\describe{
\item{NESPP3}{species code found in commercial fishing database}
\item{SVSPP}{species code found in bottom trawl survey database}
\item{NAFOSPP}{species code in nafo database}
\item{SPECIES_ITIS}{itis code}
\item{COMMON_NAME}{common name from commercial fishing database}
\item{COMNAME}{common name from bottom trawl survey fishing database}
\item{SCIENTIFIC_NAME}{scientific name of species}

}
}
\usage{
focalSpecies
}
\description{
Species name, latin name, and various related species codes
}
\keyword{datasets}
