% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskeyrun.r
\docType{package}
\name{mskeyrun}
\alias{mskeyrun}
\title{mskeyrun package}
\description{
This data package documents the decisions made when creating data products for the multispecies model
keyrun project for Georges Bank, USA. Data and methods documentation are found here.
}
\details{
\itemize{
\item Bundled data
\item Helper scrips for other file format (shp, csv)
\item Documentation of the decisions and steps taken to generate data
}

To learn more about using \code{mskeyrun}, start with the vignette: \code{browseVignettes(package="mskeyrun")} or click the index below
}
