% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realSurveyLennumcomp.r
\docType{data}
\name{realSurveyLennumcomp}
\alias{realSurveyLennumcomp}
\title{Survey length composition for the mskeyrun project based on bottom trawl survey data}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id, or Actual data}
\item{year}{year in which survey was undertaken}
\item{season}{season in which survey was undertaken}
\item{Code}{Species itis code}
\item{Name}{Common name for species}
\item{fishery}{fishery name}
\item{lenbin}{recorded 1 cm length bin}
\item{variable}{metric recorded, biomass & numbers}
\item{value}{value of the variable, summed over all tows}
\item{units}{units of the variable}

}
}
\usage{
realSurveyLennumcomp
}
\description{
Data pulled from the NEFSC bottom trawl survey database. Data pulled using
\code{\link{survdat}} package.
}
\keyword{datasets}
