% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBiolPar.R
\docType{data}
\name{simBiolPar}
\alias{simBiolPar}
\title{The Atlantis-simulated focal species biological parameters in mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{Code}{Atlantis model three letter code for functional group}
\item{Name}{Atlantis model common name for functional group}
\item{WLa}{Weight-Length equation parameter a, W = aL^b}
\item{WLb}{Weight-Length equation parameter b, W = aL^b}
}
}
\usage{
simBiolPar
}
\description{
Model name and scenario, Species code, species name, length weight parameters
}
\keyword{datasets}
