% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCatchIndex.r
\docType{data}
\name{simCatchIndex}
\alias{simCatchIndex}
\title{The Atlantis-simulated fishery catch indices for the mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{year}{year simulated fishery conducted}
\item{Code}{Atlantis model three letter code for functional group}
\item{Name}{Atlantis model common name for functional group}
\item{fishery}{simulated fishery name}
\item{variable}{catch or coefficient of variation (cv) of biomass}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
simCatchIndex
}
\description{
Simulated survey index data from Atlanits model as sampled by atlantisom
}
\keyword{datasets}
