% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFisheryLencomp.r
\docType{data}
\name{simFisheryLencomp}
\alias{simFisheryLencomp}
\title{The Atlantis-simulated fishery length composition for the mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{year}{year simulated survey conducted}
\item{Code}{Atlantis model three letter code for functional group}
\item{Name}{Atlantis model common name for functional group}
\item{fishery}{simulated fishery name}
\item{lenbin}{1 cm length bin, lower limit}
\item{variable}{number at length (Natlen)}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
simFisheryLencomp
}
\description{
Simulated fishery length composition data from Atlantis model as sampled by atlantisom
}
\keyword{datasets}
