% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFisheryWtatAgeSubannual.R
\docType{data}
\name{simFisheryWtatAgeSubannual}
\alias{simFisheryWtatAgeSubannual}
\title{The Atlantis-simulated fishery age composition for the mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{year}{year simulated fishery conducted}
\item{fishMonth}{month simulated fishery conducted}
\item{Code}{Atlantis model three letter code for functional group}
\item{Name}{Atlantis model common name for functional group}
\item{fishery}{simulated fishery name}
\item{age}{annual age of Atlantis functinal group}
\item{variable}{mean weight at age (Wtage)}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
simFisheryWtatAgeSubannual
}
\description{
Simulated fishery weight at age data from Atlanits model as sampled by atlantisom
}
\keyword{datasets}
