% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFocalSpecies.R
\docType{data}
\name{simFocalSpecies}
\alias{simFocalSpecies}
\title{The Atlantis-simulated focal species in mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{Code}{Atlantis model three letter code for functional group}
\item{Name}{Atlantis model common name for functional group}
\item{SciName}{scientific name of species}

}
}
\usage{
simFocalSpecies
}
\description{
Model name and scenario, Species name, latin name, and model species code
}
\keyword{datasets}
