% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPerCapCons.R
\docType{data}
\name{simPerCapCons}
\alias{simPerCapCons}
\title{The Atlantis-simulated age class specific per capita consumption for the mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{year}{simulation year}
\item{Code}{Atlantis model three letter code for predator group}
\item{Name}{Atlantis model common name for predator group}
\item{agecl}{age class of Atlantis functional group}
\item{variable}{total consumption (totconsagecl), total numbers (totNagecl), or annual mean per capita consumption (intakeg)}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
simPerCapCons
}
\description{
Simulated per capita consumption data by age class from Atlantis model as sampled by atlantisom
}
\keyword{datasets}
