% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStartPars.R
\docType{data}
\name{simStartPars}
\alias{simStartPars}
\title{The Atlantis-simulated focal species biological parameters in mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{Code}{Atlantis model three letter code for functional group}
\item{Name}{Atlantis model common name for functional group}
\item{agecl}{age class or true age of Atlantis functional group}
\item{lenbin}{1 cm length bin, lower limit}
\item{variable}{start year number at age class and length (Natlen), start year number at true age (Natage), average recruitment (AvgRec)}
\item{value}{value of the variable}
\item{units}{units of the variable}
}
}
\usage{
simStartPars
}
\description{
Model name and scenario, Species code, species name, length weight parameters
}
\keyword{datasets}
