% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyBottemp.R
\docType{data}
\name{simSurveyBottemp}
\alias{simSurveyBottemp}
\title{The Atlantis-simulated survey bottom temperature indices for the mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{year}{year simulated survey conducted}
\item{survey}{simulated survey name}
\item{variable}{mean annual bottom temperature degrees C}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
simSurveyBottemp
}
\description{
Simulated survey bottom temperature data from Atlanits model as sampled by atlantisom
}
\keyword{datasets}
