% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyDietcompFW.r
\docType{data}
\name{simSurveyDietcompFW}
\alias{simSurveyDietcompFW}
\title{The Atlantis-simulated survey diet composition for the mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{year}{year simulated survey conducted}
\item{Code}{Atlantis model three letter code for predator group}
\item{Name}{Atlantis model common name for predator group}
\item{survey}{simulated survey name}
\item{agecl}{age class of Atlantis functional group}
\item{prey}{Atlantis model common name for prey group}
\item{variable}{proportion in diet (dietprop)}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
simSurveyDietcompFW
}
\description{
Simulated survey diet composition data from Atlantis model as sampled by atlantisom:
Full food web beyond focal species for Rpath
}
\keyword{datasets}
