% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyIndexFW.r
\docType{data}
\name{simSurveyIndexFW}
\alias{simSurveyIndexFW}
\title{The Atlantis-simulated survey biomass indices for the mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{year}{year simulated survey conducted}
\item{Code}{Atlantis model three letter code for functional group}
\item{Name}{Atlantis model common name for functional group}
\item{survey}{simulated survey name}
\item{variable}{biomass or coefficient of variation (cv) of biomass}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
simSurveyIndexFW
}
\description{
Simulated survey index data from Atlantis model as sampled by atlantisom
Includes all structured groups as separate surveys
}
\keyword{datasets}
