% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyInfo.r
\docType{data}
\name{simSurveyInfo}
\alias{simSurveyInfo}
\title{The Atlantis-simulated survey characteristics for the mskeyrun project}
\format{
A data frame

\describe{
\item{Code}{Atlantis model three letter code for functional group}
\item{Name}{Atlantis model common name for functional group}
\item{survey}{simulated survey name}
\item{cv}{survey coefficient of variation (cv) of biomass}
\item{survMonth}{survey month of year converted from Atlantis output timestep}
\item{survArea}{survey coverage of Atlantis model polygons; "All" or "Subset"}

}
}
\usage{
simSurveyInfo
}
\description{
Simulated survey information: lookup table for species and survey names with
survey cv, timing, and general spatial coverage
}
\keyword{datasets}
