% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyLenDietcomp.r
\docType{data}
\name{simSurveyLenDietcomp}
\alias{simSurveyLenDietcomp}
\title{The Atlantis-simulated survey diet composition for the mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{year}{year simulated survey conducted}
\item{Code}{Atlantis model three letter code for predator group}
\item{Name}{Atlantis model common name for predator group}
\item{survey}{simulated survey name}
\item{pdlencm}{predator length bin (1 cm)}
\item{prey}{Atlantis model common name for prey group}
\item{variable}{proportion of prey in diet at predator length (dietprop)}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
simSurveyLenDietcomp
}
\description{
Simulated survey diet composition data from Atlantis model as sampled by atlantisom
}
\keyword{datasets}
