% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyWtatAge.R
\docType{data}
\name{simSurveyWtatAge}
\alias{simSurveyWtatAge}
\title{The Atlantis-simulated survey age composition for the mskeyrun project}
\format{
A data frame

\describe{
\item{ModSim}{Atlantis model name and simulation id}
\item{year}{year simulated survey conducted}
\item{Code}{Atlantis model three letter code for functional group}
\item{Name}{Atlantis model common name for functional group}
\item{survey}{simulated survey name}
\item{age}{annual age of Atlantis functinal group}
\item{variable}{mean weight at age (Wtage)}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
simSurveyWtatAge
}
\description{
Simulated survey weight at age data from Atlanits model as sampled by atlantisom
}
\keyword{datasets}
