% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyIndexA4.r
\docType{data}
\name{surveyIndexA4}
\alias{surveyIndexA4}
\title{The survey biomass indices for the mskeyrun project from the Albatross IV era}
\format{
A data frame

\describe{
\item{YEAR}{year survey conducted}
\item{SVSPP}{Survey Species Code see \code{focalSpecies}}
\item{SEASON}{Season the survey was conducted}
\item{variable}{stratified biomass or abundance (strat.biomass/strat.abund),
variance (biomass.var/abund.var), or standard error (biomass.SE/abund.SE)}
\item{value}{value of the variable}
\item{units}{units of the variable}

}
}
\usage{
surveyIndexA4
}
\description{
Survey data comes from Northeast Fisheries Science Center's Bottom Trawl
Survey. Indices are calculated using the \code{Survdat} R package.  This data
set covers the Albatross IV era of the bottom trawl survey from 1968 to 2008.
All appropriate calibration factors have been applied for changes in gear/vessels
over the years.  Data from the Bigelow era (>2008) are in a separate data set
(\code{surveyIndexHB}).  This data only covers the focal species of the MS Keyrun
Project.  For an expanded data set see \code{surveyIndexA4All}.
}
\keyword{datasets}
