% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyIndexA4All.r
\docType{data}
\name{surveyIndexA4All}
\alias{surveyIndexA4All}
\title{The survey biomass indices for all species within the mskeyrun project
from the Albatross IV era}
\format{
A data frame
}
\usage{
surveyIndexA4All
}
\description{
Survey data comes from Northeast Fisheries Science Center's Bottom Trawl
Survey. Indices are calculated using the \code{Survdat} R package.  This data
set covers the Albatross IV era of the bottom trawl survey from 1968 to 2008.
All appropriate calibration factors have been applied for changes in gear/vessels
over the years.  Data from the Bigelow era (>2008) are in a separate data set
(\code{surveyIndexHBAll}).  This data covers all species caught on the survey.
For the data set containing only the focal species see \code{surveyIndexA4}.
}
\seealso{
Other surveyIndex

\describe{
\item{YEAR}{year survey conducted}
\item{SVSPP}{Survey Species Code see \code{focalSpecies}}
\item{SEASON}{Season the survey was conducted}
\item{variable}{stratified biomass or abundance (strat.biomass/strat.abund),
variance (biomass.var/abund.var), or standard error (biomass.SE/abund.SE)}
\item{value}{value of the variable}
\item{units}{units of the variable}

}: 
\code{\link{surveyIndexA4}},
\code{\link{surveyIndexAll}},
\code{\link{surveyIndexHBAll}},
\code{\link{surveyIndexHB}},
\code{\link{surveyIndex}}
}
\concept{surveyIndex

\describe{
\item{YEAR}{year survey conducted}
\item{SVSPP}{Survey Species Code see \code{focalSpecies}}
\item{SEASON}{Season the survey was conducted}
\item{variable}{stratified biomass or abundance (strat.biomass/strat.abund),
variance (biomass.var/abund.var), or standard error (biomass.SE/abund.SE)}
\item{value}{value of the variable}
\item{units}{units of the variable}

}}
\keyword{datasets}
