% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheryLenSampN.r
\docType{data}
\name{fisheryLenSampN}
\alias{fisheryLenSampN}
\title{Fishery sampled number of lengths for input into hydra model}
\format{
A data frame

\describe{
\item{YEAR}{year diet data collected}
\item{SVSPP}{SVSPP code for species}
\item{lensampsize}{number of fish measured for length}
\item{ntrips}{number of unique trips that measured fish lengths}
\item{SPECIES_ITIS}{SPECIES_ITIS code for species}
\item{LongName}{Common name for species}

}
}
\usage{
fisheryLenSampN
}
\description{
Number of fish measured by year, species
for mskeyrun focal species
}
\keyword{datasets}
