% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyLenSampN.R
\docType{data}
\name{surveyLenSampN}
\alias{surveyLenSampN}
\title{Survey sampled number of lengths for input into hydra model}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1102 rows and 7 columns.
}
\usage{
surveyLenSampN
}
\description{
Number of fish measured by year, season, vessel, species
for mskeyrun focal species
}
\details{
#' @format A data frame

\describe{
\item{YEAR}{year diet data collected}
\item{SEASON}{season diet data collected}
\item{SVVESSEL}{survey vessel}
\item{SVSPP}{SVSPP code for species}
\item{lensampsize}{number of fish measured for length}
\item{SPECIES_ITIS}{SPECIES_ITIS code for species}
\item{LongName}{Common name for species}

}
}
\keyword{datasets}
